bug fixes to join implementation, added to user dir only
