+++ user/user.h
@@ -24,7 +24,6 @@ int getpid(void);
 char* sys_sbrk(int,int);
 int pause(int);
 int uptime(void);
-void join(void);
 
 // ulib.c
 int stat(const char*, struct stat*);
