+++ user/usys.pl
@@ -42,4 +42,3 @@ entry("getpid");
 entry("sbrk");
 entry("pause");
 entry("uptime");
-entry("join");
